refactor: update type and move list response type usage
